🚚 rename: 컴포넌트 이사
