+++ services/quiz/src/pages/StudyDetailPage/components/QuizItem/style.ts
