+++ services/quiz/src/pages/StudyDetailPage/components/QuizList/style.ts
