+++ services/quiz/src/pages/StudyDetailPage/components/Ranking/index.tsx
