+++ services/quiz/src/pages/StudyDetailPage/components/StudyInfo/style.ts
