+++ services/quiz/src/pages/StudyDetailPage/components/StudyProgress/index.tsx
